split elasticsearch to read and payload files
